Add room streamer to listen room data events.
