--- app/livechat/client/lib/roomManagerStream.js
@@ -1,3 +1,0 @@
-import { Meteor } from 'meteor/meteor';
-
-export const livechatRoomManager = new Meteor.Streamer('livechat-room-manager');
