+++ app/livechat/server/index.js
@@ -6,7 +6,6 @@ import './permissions';
 import '../lib/messageTypes';
 import './config';
 import './roomType';
-import './streamer';
 import './hooks/beforeGetNextAgent';
 import './hooks/externalMessage';
 import './hooks/leadCapture';
