--- app/livechat/server/streamer/index.js
@@ -1,1 +1,0 @@
-import './roomManager';
