+++ app/ui-utils/client/lib/RoomManager.js
@@ -15,6 +15,7 @@ import { Notifications } from '../../../notifications';
 import { CachedChatRoom, ChatMessage, ChatSubscription, CachedChatSubscription } from '../../../models';
 import { CachedCollectionManager } from '../../../ui-cached-collection';
 import { getConfig } from '../config';
+import { ROOM_DATA_STREAM_OBSERVER } from '../../../utils/stream/constants';
 
 import { call } from '..';
 
@@ -44,12 +45,14 @@ const onDeleteMessageBulkStream = ({ rid, ts, excludePinned, ignoreDiscussion, u
 export const RoomManager = new function() {
 	const openedRooms = {};
 	const msgStream = new Meteor.Streamer('room-messages');
+	const roomStream = new Meteor.Streamer(ROOM_DATA_STREAM_OBSERVER);
 	const onlineUsers = new ReactiveVar({});
 	const Dep = new Tracker.Dependency();
 	const Cls = class {
 		static initClass() {
 			this.prototype.openedRooms = openedRooms;
 			this.prototype.onlineUsers = onlineUsers;
+			this.prototype.roomStream = roomStream;
 			this.prototype.computation = Tracker.autorun(() => {
 				const ready = CachedChatRoom.ready.get() && mainReady.get();
 				if (ready !== true) { return; }
