+++ app/utils/lib/RoomTypeConfig.js
@@ -214,6 +214,10 @@ export class RoomTypeConfig {
 		return false;
 	}
 
+	listenStreamerEvents() {
+		return false;
+	}
+
 	/**
 	 * Returns a text which can be used in generic UIs.
 	 * @param context The role of the text in the UI-Element
