+++ app/utils/stream/constants.js
@@ -1,0 +1,1 @@
+export const ROOM_DATA_STREAM_OBSERVER = 'room-data-observer';
